Moved LottieComposition parsing to Parser classes
This required setting up a init() method on LottieComposition
to avoid exposing the fields directly outside of the lottie package.
This also removed the bodymovin min version check for text layers
because the layer parsing can no longer know the bodymovin version.
This check can be moved to render time if needed.