Fix opacity used in save layer (#1458)
contentPaint.alpha is changed multiple times without reset it back. This will lead to Utils.saveLayerCompat call with incorrect alpha.
In our case it should be 255, but if last mask was applied with 0 alpha (e.g. in applyAddMask), next draw will be transparent, because layer will be saved with 0 alpha.

Fix is to set proper alpha before saving layer.